Moved rule type into props to remove dependency cycle
I suppose you could argue this isn't the cleanest resolution, but the dependencies of these types are circular so it makes sense for them all to be in the same module.

Also removed the export of `Expression` as it is only used internally.
