+++ src/types/props.ts
@@ -1,7 +1,6 @@
 import { Color } from "./color";
 import { Font } from "./font";
 import { MacroMap } from "../redux/csState";
-import { Rule } from "./rules";
 import { WidgetActions } from "../ui/widgets/widgetActions";
 import { Border } from "./border";
 
@@ -15,3 +14,22 @@ export type GenericProp =
   | Rule[]
   | MacroMap
   | WidgetActions;
+
+interface Expression {
+  boolExp: string;
+  value: string;
+  convertedValue?: GenericProp;
+}
+
+interface PV {
+  pvName: string;
+  trigger: boolean;
+}
+
+export interface Rule {
+  name: string;
+  prop: string;
+  outExp: boolean;
+  pvs: PV[];
+  expressions: Expression[];
+}
