--- src/types/rules.ts
@@ -1,20 +1,0 @@
-import { GenericProp } from "./props";
-
-export interface Expression {
-  boolExp: string;
-  value: string;
-  convertedValue?: GenericProp;
-}
-
-interface PV {
-  pvName: string;
-  trigger: boolean;
-}
-
-export interface Rule {
-  name: string;
-  prop: string;
-  outExp: boolean;
-  pvs: PV[];
-  expressions: Expression[];
-}
